implement test_numeric
